+++ dev/src/common/core/mathFuncsFPU.inl
@@ -34,19 +34,19 @@
 /* Math functions                                                       */
 /************************************************************************/
 RED_INLINE Float MSqrt(Float f)	{	return ::sqrtf(f); }
-RED_INLINE Float MRsqrt(Float f)	{	return 1.0f/::sqrt(f); }
+RED_INLINE Float MRsqrt(Float f)	{	return 1.0f/::sqrtf(f); }
 RED_INLINE Float MAbs(Float f)	{	return ::fabsf(f); }
-RED_INLINE Float MCeil(Float f)	{	return ::ceil(f); }
-RED_INLINE Float MSin(Float f)	{	return ::sin(f); }
-RED_INLINE Float MCos(Float f)	{	return ::cos(f); }
-RED_INLINE Float MAsin(Float f)	{	RED_ASSERT( Abs<Float>(f) <= 1.f );	return ::asin(f); }
-RED_INLINE Float MAcos(Float f)	{	RED_ASSERT( Abs<Float>(f) <= 1.f ); return ::acos(f); }
-RED_INLINE Float MFloor(Float f)	{	return ::floor(f); }
-RED_INLINE Float MRound(Float f)	{	return ::floor(f + 0.5f); }
+RED_INLINE Float MCeil(Float f)	{	return ::ceilf(f); }
+RED_INLINE Float MSin(Float f)	{	return ::sinf(f); }
+RED_INLINE Float MCos(Float f)	{	return ::cosf(f); }
+RED_INLINE Float MAsin(Float f)	{	RED_ASSERT( Abs<Float>(f) <= 1.f );	return ::asinf(f); }
+RED_INLINE Float MAcos(Float f)	{	RED_ASSERT( Abs<Float>(f) <= 1.f ); return ::acosf(f); }
+RED_INLINE Float MFloor(Float f)	{	return ::floorf(f); }
+RED_INLINE Float MRound(Float f)	{	return ::floorf(f + 0.5f); }
 RED_INLINE Double MRoundD(Double d){	return ::floor(d + 0.5); }
 RED_INLINE Float MFract(Float f)  {   return f - (float)(int)f; }
-RED_INLINE Float MTan(Float f)	{	return ::tan(f); }
-RED_INLINE Float MATan2(Float y, Float x)	{ return ::atan2(y, x); }
+RED_INLINE Float MTan(Float f)	{	return ::tanf(f); }
+RED_INLINE Float MATan2(Float y, Float x)	{ return ::atan2f(y, x); }
 RED_INLINE Bool IsPow2(Int32 i)	{	return (i & ( i-1 )) == 0; }
 
 RED_INLINE float FloatSelect( float comparand, float valueGE, float valueLT )
